pyproject.toml
[tool.poetry]
name = "kafka-topic-loader"
version = "0.1.0"
description = "Creates and loads sample data into multiple kafka topics"
authors = ["Thomas Cooper <tcooper@redhat.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^1.4.7"

[tool.poetry.dev-dependencies]
ipython = "^7.11.1"
pylint = "^2.4.4"
mypy = "^0.761"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
